Move QueryProcessor out of model
