+++ irohad/model/CMakeLists.txt
@@ -29,7 +29,7 @@ target_link_libraries(sha3_hash
 add_library(model
     model_crypto_provider_impl.cpp
     impl/model_operators.cpp
-    impl/query_execution.cpp
+    execution/impl/query_execution.cpp
     )
 target_link_libraries(model
     hash
